Rework CheckWW1
