+++ Dumplings/Scanning/Scanner.cs
@@ -1,6 +1,7 @@
 ﻿using Dumplings.Analysis;
 using Dumplings.Helpers;
 using Dumplings.Rpc;
+using Dumplings.Stats;
 using Microsoft.Extensions.Caching.Memory;
 using NBitcoin;
 using NBitcoin.RPC;
@@ -137,6 +138,16 @@ namespace Dumplings.Scanning
                             // IDENTIFY WASABI COINJOINS
                             if (!isWasabi2Cj && block.Height >= Constants.FirstWasabiBlock)
                             {
+                                if (block.Height <= 584282)
+                                {
+                                }
+                                else if (block.Height <= 584306)
+                                {
+                                }
+                                else if (block.Height <= 593826)
+                                {
+                                }
+
                                 // Before Wasabi had constant coordinator addresses and different base denominations at the beginning.
                                 if (block.Height < Constants.FirstWasabiNoCoordAddressBlock)
                                 {
@@ -553,6 +564,81 @@ namespace Dumplings.Scanning
             }
 
             return denominations.OrderByDescending(x => x);
+        }
+
+        public async void DoubleCheckWW1CoinJoinsAsync(ScannerFiles loadedScannerFiles)
+        {
+            var list = new List<(YearMonth Date, VerboseTransactionInfo Tx, VerboseBlockInfo Block)>();
+            var falseList = new List<(YearMonth Date, VerboseTransactionInfo Tx, VerboseBlockInfo Block)>();
+            bool isWasabiCj = false;
+            var coinjoins = loadedScannerFiles.WasabiCoinJoins;
+
+            foreach (var tx in coinjoins)
+            {
+                var block = await Rpc.GetVerboseBlockAsync(tx.BlockInfo.BlockHash, safe: false).ConfigureAwait(false);
+
+                var indistinguishableOutputs = tx.GetIndistinguishableOutputs(includeSingle: false).ToArray();
+
+                var outputs = tx.Outputs.ToArray();
+                var inputs = tx.Inputs.Select(x => x.PrevOutput).ToArray();
+                var outputValues = outputs.Select(x => x.Value);
+                var inputValues = inputs.Select(x => x.Value);
+                var outputCount = outputs.Length;
+                var inputCount = inputs.Length;
+                (Money mostFrequentEqualOutputValue, int mostFrequentEqualOutputCount) = indistinguishableOutputs.OrderByDescending(x => x.count).First(); // Segwit only inputs.
+                var isNativeSegwitOnly = tx.Inputs.All(x => x.PrevOutput.ScriptPubKey.IsScriptType(ScriptType.P2WPKH)) && tx.Outputs.All(x => x.ScriptPubKey.IsScriptType(ScriptType.P2WPKH)); // Segwit only outputs.
+
+                if (block.Height >= Constants.FirstWasabiBlock)
+                {
+                    //if (block.Height <= 584282)
+                    //{
+                    //}
+                    //else if (block.Height <= 584306)
+                    //{
+                    //}
+                    //else if (block.Height <= 593826)
+                    //{
+                    //}
+
+                    // Before Wasabi had constant coordinator addresses and different base denominations at the beginning.
+                    if (block.Height < Constants.FirstWasabiNoCoordAddressBlock)
+                    {
+                        isWasabiCj = tx.Outputs.Any(x => Constants.WasabiCoordScripts.Contains(x.ScriptPubKey)) && indistinguishableOutputs.Any(x => x.count > 2);
+                    }
+                    else
+                    {
+                        var uniqueOutputCount = tx.GetIndistinguishableOutputs(includeSingle: true).Count(x => x.count == 1);
+                        isWasabiCj =
+                            isNativeSegwitOnly
+                            && mostFrequentEqualOutputCount >= 10 // At least 10 equal outputs.
+                            && outputValues.SequenceEqual(outputValues.OrderBy(x => x)) // Outputs are ordered ascending.
+                            && inputCount >= mostFrequentEqualOutputCount // More inptuts than most frequent equal outputs.
+                            && mostFrequentEqualOutputValue.Almost(Constants.ApproximateWasabiBaseDenomination, Constants.WasabiBaseDenominationPrecision) // The most frequent equal outputs must be almost the base denomination.
+                            && uniqueOutputCount >= 2; // It's very likely there's at least one change and at least one coord output those have unique values.
+                    }
+                }
+
+                var blockTime = tx.BlockInfo.BlockTime;
+                var blockTimeValue = blockTime.Value;
+                var yearMonth = new YearMonth(blockTimeValue.Year, blockTimeValue.Month);
+                if (isWasabiCj)
+                {
+                    list.Add((Date: yearMonth, Tx: tx, Block: block));
+                }
+                else
+                {
+                    falseList.Add((Date: yearMonth, Tx: tx, Block: block));
+                }
+            }
+            foreach (var txInfo in list)
+            {
+                Console.WriteLine($"Block: {txInfo.Block.Height}, Date:{txInfo.Date} : {txInfo.Tx.Id}");
+            }
+            Console.WriteLine("\nFALSE POSITIVES\n");
+            foreach (var txInfo in falseList)
+            {
+                Console.WriteLine($"Block: {txInfo.Block.Height}, Date:{txInfo.Date} : {txInfo.Tx.Id}");
+            }
         }
     }
 }
