+++ Dumplings/Stats/Statista.cs
@@ -1235,56 +1235,5 @@ namespace Dumplings.Stats
             }
             return myList;
         }
-
-        public void DoubleCheckWW1CoinJoinsAsync()
-        {
-            var list = new List<(YearMonth, VerboseTransactionInfo)>();
-            var falseList = new List<(YearMonth, VerboseTransactionInfo)>();
-            bool isWasabiCj = false;
-            var coinjoins = ScannerFiles.WasabiCoinJoins;
-            foreach (var tx in coinjoins)
-            {
-                var indistinguishableOutputs = tx.GetIndistinguishableOutputs(includeSingle: false).ToArray();
-
-                var outputs = tx.Outputs.ToArray();
-                var inputs = tx.Inputs.Select(x => x.PrevOutput).ToArray();
-                var outputValues = outputs.Select(x => x.Value);
-                var inputValues = inputs.Select(x => x.Value);
-                var outputCount = outputs.Length;
-                var inputCount = inputs.Length;
-                (Money mostFrequentEqualOutputValue, int mostFrequentEqualOutputCount) = indistinguishableOutputs.OrderByDescending(x => x.count).First(); // Segwit only inputs.
-                var isNativeSegwitOnly = tx.Inputs.All(x => x.PrevOutput.ScriptPubKey.IsScriptType(ScriptType.P2WPKH)) && tx.Outputs.All(x => x.ScriptPubKey.IsScriptType(ScriptType.P2WPKH)); // Segwit only outputs.
-
-                var uniqueOutputCount = tx.GetIndistinguishableOutputs(includeSingle: true).Count(x => x.count == 1);
-                isWasabiCj =
-                    isNativeSegwitOnly
-                    && mostFrequentEqualOutputCount >= 10 // At least 10 equal outputs.
-                    && outputValues.SequenceEqual(outputValues.OrderBy(x => x)) // Outputs are ordered ascending.
-                    && inputCount >= mostFrequentEqualOutputCount // More inptuts than most frequent equal outputs.
-                    && mostFrequentEqualOutputValue.Almost(Constants.ApproximateWasabiBaseDenomination, Constants.WasabiBaseDenominationPrecision) // The most frequent equal outputs must be almost the base denomination.
-                    && uniqueOutputCount >= 2; // It's very likely there's at least one change and at least one coord output those have unique values.
-
-                var blockTime = tx.BlockInfo.BlockTime;
-                var blockTimeValue = blockTime.Value;
-                var yearMonth = new YearMonth(blockTimeValue.Year, blockTimeValue.Month);
-                if (isWasabiCj)
-                {
-                    list.Add((yearMonth, tx));
-                }
-                else
-                {
-                    falseList.Add((yearMonth, tx));
-                }
-            }
-            foreach (var txInfo in list)
-            {
-                Console.WriteLine($"Block: {txInfo.Item2.BlockInfo.BlockIndex}, Date:{txInfo.Item1} : {txInfo.Item2.Id}");
-            }
-            Console.WriteLine("\nFALSE POSITIVES\n");
-            foreach (var txInfo in falseList)
-            {
-                Console.WriteLine($"{txInfo.Item1} : {txInfo.Item2.Id}");
-            }
-        }
     }
 }
